Allow strobe overlay to be configured (Pull captions out into Player)
