+++ src/renderer/style.scss
@@ -931,6 +931,12 @@ label.SimpleCheckbox {
 }
 
 .Player {
+  position: fixed;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+
   .u-button-row {
     position: absolute;
     top: 0;
